accessible/tests/mochitest/treeupdate/test_ariadialog.html

<!DOCTYPE html>
<html>

<head>
  <title>Table creation in ARIA dialog test</title>

  <link rel="stylesheet" type="text/css"
        href="chrome://mochikit/content/tests/SimpleTest/test.css" />

  <script src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>

  <script type="application/javascript"
          src="../common.js"></script>
  <script type="application/javascript"
          src="../role.js"></script>
  <script type="application/javascript"
          src="../events.js"></script>

  <script type="application/javascript">

    async function doTest() {
      // enableLogging("tree");

      let shown = waitForEvent(EVENT_SHOW, getNode("dialog"));
      getNode("dialog").style.display = "block";
      getNode("table").style.visibility = "visible";
      getNode("a").textContent = "link";
      getNode("input").value = "hello";
      getNode("input").focus();
      await shown;

      const tree = {
        role: ROLE_DIALOG,
        children: [
          {
            role: ROLE_PUSHBUTTON,
            children: [ { role: ROLE_TEXT_LEAF } ],
          },
          {
            role: ROLE_ENTRY,
          },
        ],
      };
      testAccessibleTree("dialog", tree);

      SimpleTest.finish();
    }

    SimpleTest.waitForExplicitFinish();
    addA11yLoadEvent(doTest);
  </script>
</head>
<body>

  <a target="_blank"
     title="Rework accessible tree update code"
     href="https://bugzilla.mozilla.org/show_bug.cgi?id=570275">
    Mozilla Bug 570275
  </a>

  <p id="display"></p>
  <div id="content" style="display: none"></div>
  <pre id="test">
  </pre>

  <div id="dialog" role="dialog" style="display: none;">
    <table id="table" role="presentation"
           style="display: block; position: absolute; top: 88px; left: 312.5px; z-index: 10010; visibility: hidden;">
      <tbody>
        <tr>
          <td role="presentation">
            <div role="presentation">
              <a id="a" role="button">text</a>
            </div>
            <input id="input">
          </td>
        </tr>
      </tbody>
    </table>
  </div>

  <div id="eventdump"></div>
</body>
</html>
